% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_dens_area.R
\name{color_dens_area}
\alias{color_dens_area}
\title{Color area under density function.}
\usage{
color_dens_area(density_function, color_area = c(-3, 3),
  range_x = c(-3, 3), fill_col = "steelblue", ...)
}
\arguments{
\item{density_function}{The density function to daw. Choose from e.g. http://www.stat.umn.edu/geyer/old/5101/rlook.html}

\item{color_area}{Area under the density curve to be colored.}

\item{range_x}{X-Axis range of the plot}

\item{fill_col}{Color of the specified area}

\item{...}{additional parameters for the density function}
}
\description{
Plot a density function (e.g. dnorm) and color a specified area under the curve
}
