% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpx.R
\name{dpx}
\alias{dpx}
\title{Paste Date and (optional) folder to file name}
\usage{
dpx(suffix, folder = NULL)
}
\arguments{
\item{suffix}{Single character string that is a file name that the date should be d}

\item{folder}{Give the name of an optional folder name (including "/") that is pasted in front of the filename.}
}
\description{
For easier naming of files to be saved. Adds the current date in the form of YY-MM-DD_ as a prefix to a character string. Optionally adds a folder name, too.
}
