% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplis.R
\name{duplis}
\alias{duplis}
\title{Show all duplicated values}
\usage{
duplis(x, index = T)
}
\arguments{
\item{x}{Vector to in which duplicates will be searched}

\item{index}{indicate whether to return numerical index of duplicated entries or logical vector}
}
\description{
Show all duplications in a vector and not only one.
}
\details{
Author: Holger Kirsten
}
