% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_preview.R
\name{file_preview}
\alias{file_preview}
\title{This is the title.}
\usage{
file_preview(d, mydims = 5)
}
\arguments{
\item{d}{The data.table or data.frame that is to be previewed}

\item{mydims}{The number of rows and columns to be displayed.}
}
\description{
Show the first user specified number of rows and columns of a data frame or matrix.
}
\details{
Author: Holger Kirsten
}
