% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%nin\%}
\alias{\%nin\%}
\title{%nin%}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{Vector}

\item{y}{Vector to compare against}
}
\description{
Returns a logical vector TRUE for elements of X not in Y
}
\details{
Author: Holger Kirsten
}
