% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imise_initialize.R
\name{imise_initialize}
\alias{imise_initialize}
\title{Set .libPaths when working on an IMISE server}
\usage{
imise_initialize(onServer = T, serverName = NA)
}
\arguments{
\item{onServer}{Logical. Are you working from within the IMISE network and have /net/ as your root? When F sets the basic path to /mnt/mount1/ for remote mounting}

\item{serverName}{name of the folder in rpackages with the R-packages for the server you are working on}
}
\description{
This function simply sets the correct .libPaths for the IMISE compute servers. it also returns a basic path for working from inside or via remote
}
