% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_all.R
\name{intersect_all}
\alias{intersect_all}
\title{Find Intersect}
\usage{
intersect_all(a, b, ...)
}
\arguments{
\item{a}{First vector to find intersect with the remaining input-vectors of}

\item{b}{Second vector to find intersect with the remaining input-vectors of}

\item{...}{additional vectors to find the common intersect of}
}
\description{
Find the common elemets
}
\details{
Borrowed from this thread: https://stackoverflow.com/questions/3695677/how-to-find-common-elements-from-multiple-vectors
}
