% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fdr_stats.R
\name{make_fdr_stats}
\alias{make_fdr_stats}
\title{This is the title.}
\usage{
make_fdr_stats(pval, ...)
}
\arguments{
\item{pval}{A list or vector of Pvalues to be passed to the fdrtool function}
}
\description{
A wrapper to compute FDR for a list of pvalues.
}
\details{
Author: Holger Kirsten
}
