% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_col_front.R
\name{move_col_front}
\alias{move_col_front}
\title{Move A Column To The Front}
\usage{
move_col_front(d, colname = "colname")
}
\arguments{
\item{d}{Data.table or data.frame whose columns are to be rearranged.}

\item{colname}{Character string giving the column name to be rearranged as first column}
}
\description{
Move a single column of a data frame to the very front of it, i.e. set it as the first column.
}
\details{
Author: Holger Kirsten
}
