% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newest_file.R
\name{newest_file}
\alias{newest_file}
\title{Find The Newest File.}
\usage{
newest_file(look_for = NA, subfolder = NA, directory = NA,
  print_full = F)
}
\arguments{
\item{look_for}{Character string with the name, or part of the name, of the file to look for}

\item{subfolder}{Specify the subfolder to look for the file for when location is the working directory}

\item{directory}{when Files are not in the working directory put the full path to the directory to search through here. Does not work with parameter subfolder}

\item{print_full}{If TRUE will print complete path to the file}
}
\description{
Look for a file in the working directory or a subfolder and return a character string with the newest version of the file matching the queried character string.
}
