% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newest_rprofile.R
\name{newest_rprofile}
\alias{newest_rprofile}
\title{Newest Rprofile by Holger Kirsten.}
\usage{
newest_rprofile()
}
\description{
This function looks for and loads the newest .RProfile file by Holger when executed in the IMISE network environment.
}
\details{
The loaded RProfile contains helper functions.
}
