% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformate_for_matrixEQTL.R
\name{reformate_for_matrixEQTL}
\alias{reformate_for_matrixEQTL}
\title{Reformate For MatrixEQTL.}
\usage{
reformate_for_matrixEQTL(metab4_m_po, idvar = "pid")
}
\arguments{
\item{metab4_m_po}{The data.table to reformat}

\item{idvar}{The column name specifying the probe IDs}
}
\description{
Function to transpose data for processing in MatrixEQTL.
}
\details{
Author: Holger Kirsten
}
