% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_csv_carl.R
\name{save_csv_carl}
\alias{save_csv_carl}
\title{Custom save to .csv function}
\usage{
save_csv_carl(file = NA, file_name = NA, subfolder = NA,
  create_subfolder = F)
}
\arguments{
\item{file}{The data.table or data.frame to be saved.}

\item{file_name}{Character string that specifies the name the saved file should have
The date of creation and the .csv ending are added automatically}

\item{subfolder}{A character string without "/" giving the subfolder the file shall be saved in.}

\item{create_subfolder}{Given a subfolder, setting this to TRUE will create a new directory with the name given in subfolder and will stop if set to FALSE}
}
\description{
This function saves data.tables or data.frames as .csv in the root working directory or a specified subfolder.
Additionally the current date is automatically included in the file name.
}
