% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_NA.R
\name{show_NA}
\alias{show_NA}
\title{Show Missing Values.}
\usage{
show_NA(x)
}
\arguments{
\item{x}{Data.table or data.frame to look for missing values in}
}
\description{
Show all NAs per column in a data.table or data.frame
}
\details{
Author: Holger Kirsten
}
