% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract.R
\name{str_extract}
\alias{str_extract}
\title{Extract strings from a vector of strings}
\usage{
str_extract(string, extract)
}
\arguments{
\item{string}{character string or vector containing expressions to look for}

\item{extract}{Regular expression to look for in string}
}
\description{
Using only base regex, return a matching string based on the posssibilites
entered in extract. Was intended to find a matching string when several
possibilites exist. For example, when searching str_extract(c("Case_Two",
"Case_One"), "One|Two"), the function will return the found regex in the
order they were found in the vector of strings, i.e. c("Two", "One").
}
