% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_carl.R
\name{theme_carl}
\alias{theme_carl}
\title{A custom ggplot2 theme.}
\usage{
theme_carl()
}
\description{
This is a ggplot2 theme for making minimalistic plots with.
}
\details{
A small adjustment made to the theme_minimal() as was suggested here:
https://sakaluk.wordpress.com/2016/02/16/7-make-it-pretty-plots-for-meta-analysis/
}
