% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_2.R
\name{venn2}
\alias{venn2}
\title{2-Way Venn Diagram}
\usage{
venn2(x1, y1, mytitle = "2-Way Venn Diagram", mylabels = NA, plotte = T)
}
\arguments{
\item{x1}{first vector to be matched}

\item{y1}{second vector to be matched}

\item{mytitle}{Title of the plot}

\item{mylabels}{Labels for plotting}

\item{plotte}{...}
}
\description{
Plot a non-proportional 2-Way Venn Diagram.
This function is from the package sytemPipeR found here:
http://faculty.ucr.edu/~tgirke/Documents/R_BioCond/My_R_Scripts/vennDia.R
or
https://bioconductor.org/packages/release/bioc/html/systemPipeR.html
}
\details{
Please check the systemPipeR documentation for further details
}
