% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_3.R
\name{venn3}
\alias{venn3}
\title{3-Way Venn Diagram}
\usage{
venn3(x1, y1, z1, mytitle = "3-Way Venn Diagram", mylabels = NA,
  plotte = T)
}
\arguments{
\item{x1}{first vector to be matched}

\item{y1}{second vector to be matched}

\item{z1}{third vector to be matched}

\item{mytitle}{Title of the plot}

\item{mylabels}{Labels for plotting}

\item{plotte}{...}
}
\description{
Plot a non-proportional 3-Way Venn Diagram.
This function is from the package sytemPipeR found here:
http://faculty.ucr.edu/~tgirke/Documents/R_BioCond/My_R_Scripts/vennDia.R
#' or
https://bioconductor.org/packages/release/bioc/html/systemPipeR.html
}
\details{
Please check the systemPipeR documentation for further details
}
