% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_diagram.R
\name{venndiagram}
\alias{venndiagram}
\title{Pretty Venn Diagrams}
\usage{
venndiagram(x = x, y = y, z = z, w = w, unique = T,
  title = "Venn Diagram", labels = c("x", "y", "z", "w"), lines = 1,
  lcol = 1, tcol = 1, diacol = 1, plot = T, type = "3",
  printsub = TRUE, ...)
}
\description{
Plot a non-proportional 2-Way, 3-Way or 4-Way Venn Diagram.
This function is from the package sytemPipeR found here:
http://faculty.ucr.edu/~tgirke/Documents/R_BioCond/My_R_Scripts/vennDia.R
or
https://bioconductor.org/packages/release/bioc/html/systemPipeR.html
}
\details{
Please check the systemPipeR documentation for further details
}
