% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_csv_carl.R
\name{save_csv_carl}
\alias{save_csv_carl}
\title{Custom save to .csv function}
\usage{
save_csv_carl(file = NA, file_name = NA, subfolder = NA,
  create_subfolder = F, sep = ",", quote = F)
}
\arguments{
\item{file}{The data.table or data.frame to be saved.}

\item{file_name}{Character string that specifies the name the saved file should have
The date of creation and the .csv ending are added automatically}

\item{subfolder}{A character string without "/" giving the subfolder the file shall be saved in.}

\item{create_subfolder}{Given a subfolder, setting this to TRUE will create a new directory with the name given in subfolder and will stop if set to FALSE}

\item{sep}{the field separator string. Values within each row of x are separated by this string.}

\item{quote}{a logical value (TRUE or FALSE) or a numeric vector. If TRUE, any character or factor columns will be surrounded by double quotes. If a numeric vector, its elements are taken as the indices of columns to quote. In both cases, row and column names are quoted if they are written. If FALSE, nothing is quoted.}
}
\description{
This function saves data.tables or data.frames as .csv in the root working directory or a specified subfolder.
Additionally the current date is automatically included in the file name.
}
